+++ packages/cubejs-mongobi-driver/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/statsbotco/cubejs-client/compare/v0.5.0...v0.5.1) (2019-04-02)
+
+
+### Features
+
+* BigQuery driver ([654edac](https://github.com/statsbotco/cubejs-client/commit/654edac))
+
+
+
+
+
 # [0.5.0](https://github.com/statsbotco/cubejs-client/compare/v0.4.6...v0.5.0) (2019-04-01)
 
 **Note:** Version bump only for package @cubejs-backend/mongobi-driver