+++ packages/cubejs-playground/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/statsbotco/cubejs-client/compare/v0.5.0...v0.5.1) (2019-04-02)
+
+**Note:** Version bump only for package @cubejs-client/playground
+
+
+
+
+
 # [0.5.0](https://github.com/statsbotco/cubejs-client/compare/v0.4.6...v0.5.0) (2019-04-01)
 
 **Note:** Version bump only for package @cubejs-client/playground