+++ packages/cubejs-vue/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/statsbotco/cubejs-client/compare/v0.5.0...v0.5.1) (2019-04-02)
+
+
+### Features
+
+* Vue package improvements and docs ([fc38e69](https://github.com/statsbotco/cubejs-client/commit/fc38e69)), closes [#68](https://github.com/statsbotco/cubejs-client/issues/68)
+
+
+
+
+
 # [0.5.0](https://github.com/statsbotco/cubejs-client/compare/v0.4.6...v0.5.0) (2019-04-01)